+++ test/dynamic_color_builder_test.dart
@@ -1,4 +1,5 @@
 import 'package:dynamic_color/dynamic_color.dart';
+import 'package:dynamic_color/samples.dart';
 import 'package:dynamic_color/test_utils.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -7,7 +8,7 @@ void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
 
   testWidgets('DynamicColorBuilder', (WidgetTester tester) async {
-    DynamicColorTestingUtils.setMockDynamicColors(sampleOfCorePalette);
+    DynamicColorTestingUtils.setMockDynamicColors(SampleCorePalettes.green);
 
     const containerKey = Key('myContainer');
